+++ config.inc.php
@@ -130,6 +130,9 @@ $REX['ADDON']['seo42']['settings'] = array(
 	)
 );
 
+// backup default settings for debug mode 
+$REX['ADDON']['seo42']['settings_backup'] = $REX['ADDON']['seo42']['settings'];
+
 // overwrite default settings with user settings
 seo42_utils::includeSettingsFile();