+++ pages/help/debug.inc.php
@@ -1,12 +1,6 @@
 <?php
 
-$REX['REDAXO'] = false;
-$REX['ADDON']['seo42']['settings']['include_query_params'] = false;
-
 $debugOut = seo42::getDebugInfo($REX['START_ARTICLE_ID']);
-
-$REX['REDAXO'] = true;
-$REX['ADDON']['seo42']['settings']['include_query_params'] = true;
 
 if ($debugOut) {
 	echo $debugOut;